[FEAT] Pull out components
