+++ src/features/world/ui/cornMaze/MazeHud.tsx
@@ -1,13 +1,8 @@
 import { SUNNYSIDE } from "assets/sunnyside";
-import { Button } from "components/ui/Button";
 import { CountdownLabel } from "components/ui/CountdownLabel";
-import { NPC_WEARABLES } from "lib/npcs";
 import React, { useContext, useEffect } from "react";
 import { Modal } from "react-bootstrap";
 import { useNavigate } from "react-router-dom";
-
-import crowWithoutShadow from "assets/decorations/crow_without_shadow.png";
-import crowFeather from "assets/decorations/crow_feather_large.png";
 import classNames from "classnames";
 import { Context } from "features/game/GameProvider";
 import { MachineState as GameMachineState } from "features/game/lib/gameMachine";
@@ -15,17 +10,23 @@ import { useInterpret, useSelector } from "@xstate/react";
 import { calculateFeathersEarned } from "features/game/events/landExpansion/attemptMaze";
 import { getSeasonWeek } from "lib/utils/getSeasonWeek";
 import { MazeMetadata, WitchesEve } from "features/game/types/game";
-import { Panel } from "components/ui/Panel";
 import {
   MachineInterpreter,
   MachineState,
   cornMazeMachine,
-} from "../lib/cornmazeMachine";
+} from "../../lib/cornmazeMachine";
+import { LosingModalContent } from "./LosingModalContent";
+import { PausedHighScoreModalContent } from "./PausedHighScoreModalContent";
+import { PausedLowScoreModalContent } from "./PausedLowScoreModalContent";
+import { TipsModalContent } from "./TipsModalContent";
+import { WinningModalContent } from "./WinningModalContent";
+
+import crowWithoutShadow from "assets/decorations/crow_without_shadow.png";
 
 type Listener = {
   collectCrow: (id: string) => void;
   hit: () => void;
-  sceneLoaded: (crowCount: number) => void;
+  sceneLoaded: () => void;
   handlePortalHit: () => void;
 };
 class MazeManager {
@@ -43,9 +44,9 @@ class MazeManager {
     }
   }
 
-  public sceneLoaded(crowCount: number) {
+  public sceneLoaded() {
     if (this.listener) {
-      this.listener.sceneLoaded(crowCount);
+      this.listener.sceneLoaded();
     }
   }
 
@@ -74,7 +75,6 @@ const _health = (state: MachineState) => state.context.health;
 const _timeElapsed = (state: MachineState) => state.context.timeElapsed;
 const _startedAt = (state: MachineState) => state.context.startedAt;
 
-const _playing = (state: MachineState) => state.matches("playing");
 const _paused = (state: MachineState) => state.matches("paused");
 const _wonGame = (state: MachineState) => state.matches("wonGame");
 const _lostGame = (state: MachineState) => state.matches("lostGame");
@@ -89,17 +89,6 @@ export const MazeHud: React.FC = () => {
   const { claimedFeathers, highestScore } = witchesEve?.maze[
     currentWeek
   ] as MazeMetadata;
-
-  // const [score, setScore] = useState(0);
-  // const [health, setHealth] = useState(3);
-  // const [totalLostCrows, setTotalLostCrowCount] = useState(0);
-  // const [gameOver, setGameOver] = useState<"won" | "lost">();
-  // const [sceneLoaded, setSceneLoaded] = useState(false);
-  // const [showTips, setShowTips] = useState(false);
-  // const [startedAt, setStartedAt] = useState<number>(0);
-  // const [timeElapsed, setTimeElapsed] = useState<number>(0);
-  // const [paused, setPaused] = useState(false);
-  // const [pausedAt, setPausedAt] = useState(0);
 
   const navigate = useNavigate();
 
@@ -129,43 +118,26 @@ export const MazeHud: React.FC = () => {
   useEffect(() => {
     mazeManager.listen({
       collectCrow: () => {
-        // setScore((s) => s + 1);
         cornMazeService.send("COLLECT_CROW");
       },
       hit: () => {
-        // setHealth((h) => h - 1);
         cornMazeService.send("HIT");
       },
-      sceneLoaded: (crowCount: number) => {
-        // setTotalLostCrowCount(crowCount);
+      sceneLoaded: () => {
         cornMazeService.send("SCENE_LOADED");
       },
       handlePortalHit: () => {
-        // setGameOver("won");
         cornMazeService.send("PORTAL_HIT");
       },
     });
   }, []);
 
   const handleStart = () => {
-    // setStartedAt(Date.now());
     cornMazeService.send("START_GAME");
   };
 
   const handleResume = () => {
-    // setPaused(false);
-    // setTimeElapsed((e) => e + Date.now() - pausedAt);
     cornMazeService.send("RESUME_GAME");
-  };
-
-  const getFeathersEarned = () => {
-    const weeklyMazeData = witchesEve?.maze[currentWeek];
-
-    if (!weeklyMazeData) return 0;
-
-    const { claimedFeathers } = weeklyMazeData;
-
-    return calculateFeathersEarned(weeklyLostCrowCount, score, claimedFeathers);
   };
 
   const handleMazeComplete = () => {
@@ -187,6 +159,16 @@ export const MazeHud: React.FC = () => {
   const handleShowTips = () => {
     // setShowTips(true);
     cornMazeService.send("SHOW_TIPS");
+  };
+
+  const getFeathersEarned = () => {
+    const weeklyMazeData = witchesEve?.maze[currentWeek];
+
+    if (!weeklyMazeData) return 0;
+
+    const { claimedFeathers } = weeklyMazeData;
+
+    return calculateFeathersEarned(weeklyLostCrowCount, score, claimedFeathers);
   };
 
   if (!sceneLoaded) return null;
@@ -296,153 +278,5 @@ export const MazeHud: React.FC = () => {
         />
       </Modal>
     </>
-  );
-};
-
-const TipsModalContent: React.FC<{
-  gameActive: boolean;
-  onStart: () => void;
-  onResume: () => void;
-}> = ({ gameActive, onStart, onResume }) => {
-  return (
-    <Panel
-      bumpkinParts={{
-        ...NPC_WEARABLES.luna,
-        body: "Light Brown Worried Farmer Potion",
-      }}
-    >
-      <>
-        <div className="p-1 pt-2 space-y-2 mb-2">
-          <div className="space-y-2 flex flex-col">
-            <div className="flex items-center space-x-2">
-              <img src={crowWithoutShadow} alt="Corn" className="w-6" />
-              <p>Collect all the missing crows.</p>
-            </div>
-            <div className="flex items-center space-x-2">
-              <img src={SUNNYSIDE.icons.heart} alt="Health" className="w-6" />
-              <p>Avoid all the enemies.</p>
-            </div>
-            <div className="flex items-center space-x-2">
-              <img
-                src={SUNNYSIDE.icons.stopwatch}
-                alt="timer"
-                className="h-6"
-              />
-              <p>Make it back to the portal before the time runs out!</p>
-            </div>
-            <div className="flex items-center space-x-2">
-              <img src={crowFeather} alt="feather" className="h-6" />
-              <p>{`Earn up to 100 feathers each week.`}</p>
-            </div>
-          </div>
-        </div>
-        {!gameActive && <Button onClick={onStart}>{`Let's Go!`}</Button>}
-        {gameActive && <Button onClick={onResume}>{`Got it`}</Button>}
-      </>
-    </Panel>
-  );
-};
-
-const LosingModalContent: React.FC<{
-  timeRemaining: number;
-  onClick: () => void;
-}> = ({ timeRemaining, onClick }) => {
-  return (
-    <Panel
-      bumpkinParts={{
-        ...NPC_WEARABLES.luna,
-        body: "Light Brown Worried Farmer Potion",
-      }}
-    >
-      <div className="p-1 space-y-2 mb-2 flex flex-col">
-        <p>
-          {`Oh no ${
-            timeRemaining === 0 ? "times up" : ""
-          }! My poor crows! It seems you have been outwitted by the cunning
-          enemies. For now, you shall return to whence you came.`}
-        </p>
-        <p>
-          The magical corn maze bids you farewell, brave adventurer. Until next
-          time!
-        </p>
-      </div>
-      <Button onClick={onClick}>Back to Plaza</Button>
-    </Panel>
-  );
-};
-
-const WinningModalContent: React.FC<{
-  feathersEarned: number;
-  claimedFeathers: number;
-  onClick: () => void;
-}> = ({ feathersEarned, claimedFeathers, onClick }) => {
-  return (
-    <Panel bumpkinParts={NPC_WEARABLES.luna}>
-      <div className="p-1 space-y-2 mb-2 flex flex-col">
-        <p>
-          {`Ah, you've done it! You found all my mischievous crows hidden in
-      the corn maze! I am absolutely delighted! `}
-        </p>
-        <p>
-          {`I bestow upon you ${feathersEarned} valuable crow ${
-            feathersEarned > 1 ? "feathers" : "feather"
-          }, shimmering
-          with magic to bless your future journeys.`}
-        </p>
-        <p className="text-xxs pb-1 italic">{`Feathers earned this week: ${claimedFeathers}/100`}</p>
-      </div>
-      <Button onClick={onClick}>Claim Crow Feathers</Button>
-    </Panel>
-  );
-};
-
-const PausedHighScoreModalContent: React.FC<{
-  score: number;
-  feathersEarned: number;
-  claimedFeathers: number;
-  onContinue: () => void;
-  onEnd: () => void;
-}> = ({ score, feathersEarned, claimedFeathers, onContinue, onEnd }) => {
-  return (
-    <Panel bumpkinParts={NPC_WEARABLES.luna}>
-      <div className="p-1 space-y-2 mb-2 flex flex-col">
-        <p>
-          {`You found ${score} of my mischievous crows hidden in
-      the corn maze! For your efforts, I will bestow upon you ${feathersEarned} valuable crow feathers.`}
-        </p>
-        <p>Are you sure you want to end now?</p>
-        <p className="text-xxs italic">{`Feathers earned this week: ${claimedFeathers}/100`}</p>
-      </div>
-      <div className="flex flex-col-reverse space-y-1 space-y-reverse md:flex-row md:space-y-0 md:space-x-1">
-        <Button onClick={onContinue}>Keep Playing</Button>
-        <Button onClick={onEnd}>Claim Crow Feathers</Button>
-      </div>
-    </Panel>
-  );
-};
-
-const PausedLowScoreModalContent: React.FC<{
-  highestScore: number;
-  claimedFeathers: number;
-  onContinue: () => void;
-  onEnd: () => void;
-}> = ({ highestScore, claimedFeathers, onContinue, onEnd }) => {
-  return (
-    <Panel
-      bumpkinParts={{
-        ...NPC_WEARABLES.luna,
-        body: "Light Brown Worried Farmer Potion",
-      }}
-    >
-      <div className="p-1 space-y-2 mb-2 flex flex-col">
-        <p>{`Oh no! Last time you found ${highestScore} crows! You will need to find more than that if you want more feathers from me!`}</p>
-        <p>Are you sure you want to end now?</p>
-        <p className="text-xxs italic">{`Feathers earned this week: ${claimedFeathers}/100`}</p>
-      </div>
-      <div className="flex flex-col-reverse space-y-1 space-y-reverse md:flex-row md:space-y-0 md:space-x-1">
-        <Button onClick={onEnd}>Back to Plaza</Button>
-        <Button onClick={onContinue}>Keep Playing</Button>
-      </div>
-    </Panel>
   );
 };
